refactor: 클래스 분리, 패키지 정리
