+++ src/main/java/subway/function/line/management/PrintAllLinesList.java
@@ -1,0 +1,11 @@
+package subway.function.line.management;
+
+import subway.domain.line.LineRepository;
+import subway.function.line.printer.LineManagementPrinter;
+
+public class PrintAllLinesList {
+    public static void printAllLinesList() {
+        LineManagementPrinter.printLineListTitle();
+        LineRepository.printAllLines();
+    }
+}
