--- src/main/java/subway/function/section/SectionManagement.java
@@ -1,97 +1,0 @@
-package subway.function.section;
-
-import java.util.Scanner;
-import subway.common.print.info.CommonInfoPrinter;
-import subway.common.validator.CommonValidator;
-import subway.domain.LineStationMappingRepository;
-import subway.domain.line.Line;
-import subway.domain.station.Station;
-import subway.function.section.printer.PrintSectionManagementScreen;
-import subway.function.section.printer.SectionManagementPrinter;
-
-public class SectionManagement {
-    public static void start(Scanner scanner) {
-        while (true) {
-            SectionManagementType type = printAndGetUserSelectionInput(scanner);
-            if (type == SectionManagementType.GO_BACK) {
-                return;
-            }
-            SectionManagementTypeResolver.resolveSectionManagement(type, scanner);
-        }
-    }
-
-    private static SectionManagementType printAndGetUserSelectionInput(Scanner scanner) {
-        PrintSectionManagementScreen.printSectionManagementScreen();
-        CommonInfoPrinter.printUserFunctionSelectionMessage();
-        String userInput = scanner.nextLine();
-        try {
-            CommonValidator
-                .validateIsCorrectSelectionInput(CommonValidator.SELECTION_INPUT_PATTERN_12B,
-                    userInput);
-        } catch (IllegalArgumentException e) {
-            return SectionManagementType.ERROR;
-        }
-        return SectionManagementTypeResolver.getSectionManagementType(userInput);
-    }
-
-    public static void registerNewSection(Scanner scanner) throws IllegalArgumentException {
-        Line lineToRegisterSection = getLineToRegisterNewSection(scanner);
-        Station stationToRegisterSection = getStationToRegisterSection(scanner,
-            lineToRegisterSection);
-        int orderToRegisterSection = getOrderToRegisterSection(scanner, lineToRegisterSection);
-        LineStationMappingRepository
-            .registerNewSectionByLineAndStation(lineToRegisterSection, stationToRegisterSection,
-                orderToRegisterSection);
-        SectionManagementPrinter.printSectionRegistrationSuccessMessage();
-    }
-
-    private static Line getLineToRegisterNewSection(Scanner scanner) {
-        SectionManagementPrinter.printLineNameToRegisterSectionInputMessage();
-        String lineNameToRegisterSection = scanner.nextLine();
-        return CommonValidator.validateIsLineNameExists(lineNameToRegisterSection);
-    }
-
-    private static Station getStationToRegisterSection(Scanner scanner,
-        Line lineToRegisterSection) {
-        SectionManagementPrinter.printStationNameToRegisterSectionInputMessage();
-        String stationNameToRegisterSection = scanner.nextLine();
-        Station stationToRegisterSection = CommonValidator
-            .validateIsStationNameExists(stationNameToRegisterSection);
-        return SectionManagementValidator
-            .validateStationNotInThatLine(stationToRegisterSection, lineToRegisterSection);
-    }
-
-    private static int getOrderToRegisterSection(Scanner scanner,
-        Line lineToRegisterSection) {
-        SectionManagementPrinter.printOrderToRegisterInputMessage();
-        String orderStrToRegisterSection = scanner.nextLine();
-        return SectionManagementValidator
-            .validateOrderStrToRegisterSection(orderStrToRegisterSection, lineToRegisterSection);
-    }
-
-    public static void deleteSection(Scanner scanner) throws IllegalArgumentException {
-        Line lineToDeleteSection = getLineToDeleteSection(scanner);
-        Station stationToDeleteSection = getStationToDeleteSection(scanner, lineToDeleteSection);
-        LineStationMappingRepository
-            .deleteSection(lineToDeleteSection, stationToDeleteSection);
-        SectionManagementPrinter.printSectionDeleteSuccessMessage();
-    }
-
-    private static Line getLineToDeleteSection(Scanner scanner) {
-        SectionManagementPrinter.printLineNameToDeleteSectionInputMessage();
-        String lineNameToDeleteSection = scanner.nextLine();
-        Line lineToDeleteSection = CommonValidator
-            .validateIsLineNameExists(lineNameToDeleteSection);
-        return SectionManagementValidator.validateLineHasMoreThanMinSizeToDeleteSection(lineToDeleteSection);
-    }
-
-    private static Station getStationToDeleteSection(Scanner scanner,
-        Line lineToRegisterSection) {
-        SectionManagementPrinter.printStationNameToDeleteSectionInputMessage();
-        String stationNameToDeleteSection = scanner.nextLine();
-        Station stationToDeleteSection = CommonValidator
-            .validateIsStationNameExists(stationNameToDeleteSection);
-        return SectionManagementValidator
-            .validateStationInThatLine(stationToDeleteSection, lineToRegisterSection);
-    }
-}
