+++ src/main/java/subway/function/station/management/PrintAllStationsList.java
@@ -1,0 +1,9 @@
+package subway.function.station.management;
+
+import subway.domain.station.StationRepository;
+
+public class PrintAllStationsList {
+    public static void printAllStationsList() {
+        StationRepository.printAll();
+    }
+}
