+++ src/main/java/subway/function/station/type/StationManagementTypeResolver.java
@@ -1,10 +1,12 @@
-package subway.function.station;
+package subway.function.station.type;
 
 import java.util.Scanner;
-import subway.main.UserSelections;
+import subway.function.station.management.DeleteStation;
+import subway.function.station.management.PrintAllStationsList;
+import subway.function.station.management.RegisterNewStation;
+import subway.main.type.UserSelections;
 
 public class StationManagementTypeResolver {
-
     public static StationManagementSelectionType getStationManagementSelectionType(
         String userInput) {
         if (userInput.equals(UserSelections.FIRST)) {
@@ -23,13 +25,13 @@ public class StationManagementTypeResolver {
         Scanner scanner) {
         try {
             if (type == StationManagementSelectionType.STATION_REGISTRATION) {
-                StationManagement.registerNewStation(scanner);
+                RegisterNewStation.registerNewStation(scanner);
             }
             if (type == StationManagementSelectionType.STATION_DELETE) {
-                StationManagement.deleteStation(scanner);
+                DeleteStation.deleteStation(scanner);
             }
             if (type == StationManagementSelectionType.STATION_PRINT_ALL) {
-                StationManagement.printAllStations();
+                PrintAllStationsList.printAllStationsList();
             }
         } catch (IllegalArgumentException ignored) {
         }
