+++ src/main/java/subway/main/SubwayMapProgram.java
@@ -3,6 +3,9 @@ package subway.main;
 import java.util.Scanner;
 import subway.common.validator.CommonValidator;
 import subway.common.print.info.CommonInfoPrinter;
+import subway.main.print.PrintMainScreen;
+import subway.main.type.MainSelectionType;
+import subway.main.type.MainTypeResolver;
 
 public class SubwayMapProgram {
